Try to make clearTags() more atomic
To avoid  tagged items in the cache that are not associated to a tag-list anymore, the deletion of the item should be as close to the deletion from the list as possible and not in to seperate steps to first remove all keys from the lists and afterwards remove all keys.
Also as the clearTags() operation is not atomic the cleared list should not be removed as in the meantime new keys might have been added to the list.